Move `config/task-config`
We move the folder `config/task-config` to `scripts`, renaming it to `migrate-task-config` to match the calling script
