+++ package.json
@@ -11,8 +11,8 @@
   "license": "OGL-UK-3.0",
   "scripts": {
     "test": "DATABASE_URL=postgres://water_user:password@localhost:5432/wabs_test LAZY_REDIS=1 lab",
-    "migrate": "node scripts/create-schema && db-migrate up --verbose && node scripts/migrate-task-config",
-    "migrate:local-test": "DATABASE_URL=postgres://water_user:password@localhost:5432/wabs_test node scripts/create-schema && DATABASE_URL=postgres://water_user:password@localhost:5432/wabs_test db-migrate up --verbose && DATABASE_URL=postgres://water_user:password@localhost:5432/wabs_test node scripts/migrate-task-config",
+    "migrate": "node scripts/create-schema && db-migrate up --verbose && node scripts/migrate-task-config/migrate-task-config.js",
+    "migrate:local-test": "DATABASE_URL=postgres://water_user:password@localhost:5432/wabs_test node scripts/create-schema && DATABASE_URL=postgres://water_user:password@localhost:5432/wabs_test db-migrate up --verbose && DATABASE_URL=postgres://water_user:password@localhost:5432/wabs_test node scripts/migrate-task-config/migrate-task-config.js",
     "migrate:down": "db-migrate down --verbose",
     "migrate:create": "db-migrate create --sql-file --",
     "lint": "standard",
