+++ scripts/migrate-task-config/build-config-object.js
