+++ scripts/migrate-task-config/hof-stop/config.json
